more decomp of menus and combining files
